requirements.txt
streamlit
requests
requests-toolbelt
python-dotenv
groq